추가 Entity Class
JPA 프로젝트를 생성해서 Entity를 자동생성하여 가져옴

@JoinColumn 에서 insert, update 관련 오류가 발생하여 insertable,
updatable을 false 로 설정하였음.

@OneToMany와 @ManyToOne 관계에서 infinite Loop가 발생하였음.

@ManyToOne 에는 @JsonManagedReference
@OneToMany 에는 @JsonBackReference

를 선언 함으로써 Infinite Loop 해소

+++ src/main/java/com/karljeong/fourtysix/application/admin/codeGroup/service/CodeGroupService.java
@@ -1,5 +1,6 @@
 package com.karljeong.fourtysix.application.admin.codeGroup.service;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -36,7 +37,7 @@ public class CodeGroupService {
 	}
 
 	public TbComCodeGroup create(TbComCodeGroup tbComCodeGroup) {
-		tbComCodeGroup.setCreateUserId(Long.valueOf(11111));
+		tbComCodeGroup.setCreateUserId(BigInteger.valueOf(11111));
 		TbComCodeGroup save = tbComCodeGroupRepository.save(tbComCodeGroup);
 		return save;
 	}
